pyproject.toml
[project]
name = "bipocsf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "snowflake>=1.8.0",
    "sseclient-py>=1.8.0",
    "streamlit>=1.50.0",
    "ty>=0.0.1a21",
    "uvicorn>=0.37.0",
]
